backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
langchain>=1.1.2
langchain-openai>=1.1.0
langchain-anthropic>=1.2.0
langchain-google-genai>=3.2.0
pydantic==2.9.0
python-dotenv==1.0.1